refactor(haptics): remove runtime require barrel, use file-based platform split
Replace `Platform.OS === 'web' ? require('./haptics.web') : require('./haptics.native')`
barrel with direct file-based platform resolution (index.ts for native,
index.web.ts for web). Metro picks the correct variant automatically.

Part of LL-014. Refs #117
