+++ src/libs/haptics/index.ts
@@ -1,5 +1,14 @@
-import { Platform } from "react-native";
+import * as ExpoHaptics from "expo-haptics";
 
-const Haptics = Platform.OS === "web" ? require("./haptics.web").default : require("./haptics.native").default;
+export const ImpactFeedbackStyle = ExpoHaptics.ImpactFeedbackStyle;
+export const NotificationFeedbackType = ExpoHaptics.NotificationFeedbackType;
+
+const Haptics = {
+  selectionAsync: (): Promise<void> => ExpoHaptics.selectionAsync(),
+  impactAsync: (style: ExpoHaptics.ImpactFeedbackStyle): Promise<void> => ExpoHaptics.impactAsync(style),
+  notificationAsync: (type: ExpoHaptics.NotificationFeedbackType): Promise<void> => ExpoHaptics.notificationAsync(type),
+  ImpactFeedbackStyle,
+  NotificationFeedbackType,
+};
 
 export default Haptics;
